tt/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-mlops-platform"
version = "1.0.0"
description = "Enterprise MLOps Platform with FastAPI, MLflow, Ray, DVC, Feast, and comprehensive monitoring"
readme = "README.md"
authors = [
    {name = "Enterprise ML Team", email = "mlops@enterprise.com"}
]
license = {text = "MIT"}
keywords = ["mlops", "machine-learning", "fastapi", "mlflow", "ray", "dvc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "SQLAlchemy==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.12.1",
    "psycopg2-binary==2.9.9",
    "mlflow==3.10.1",
    "ray[default]==2.54.1",
    "dvc[s3]==3.66.1",
    "great-expectations==1.15.2",
    "whylogs==1.6.4",
    "evidently==0.7.21",
    "arize-phoenix==13.22.2",
    "feast==0.61.0",
    "boto3==1.34.0",
    "aiobotocore==2.11.1",
    "neo4j==5.19.0",
    "pymysql==1.1.0",
    "redis==5.0.1",
    "celery==5.3.4",
    "flower==2.0.1",
    "prometheus-client==0.19.0",
    "starlette-prometheus==0.8.0",
    "pyyaml==6.0.1",
    "pandas==2.1.4",
    "numpy==1.24.4",
    "scikit-learn==1.3.2",
    "xgboost==2.0.2",
    "lightgbm==4.1.0",
    "torch==2.1.2",
    "transformers==4.36.2",
    "datasets==2.15.0",
    "polars==0.20.7",
    "duckdb==0.9.2",
    "prefect==2.16.1",
    "kombu==5.3.5",
    "click==8.1.7",
    "jinja2==3.1.3",
    "httpx==0.25.2",
    "tenacity==8.2.3",
    "cryptography==41.0.7",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-multipart==0.0.6",
    "email-validator==2.1.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-xdist==3.5.0",
    "freezegun==1.2.2",
    "responses==0.24.1",
    "black==23.12.1",
    "isort==5.13.2",
    "flake8==6.1.0",
    "mypy==1.8.0",
    "pre-commit==3.6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter==1.0.0",
    "jupyterlab==4.0.10",
    "ipython==8.19.0",
    "ipykernel==6.28.0",
    "matplotlib==3.8.2",
    "seaborn==0.13.0",
    "plotly==5.18.0",
    "streamlit==1.29.0",
    "gradio==4.13.0",
    "tensorboard==2.15.1",
    "networkx==3.2.1",
    "graphviz==0.20.1",
]
gpu = [
    "torch==2.1.2+cu121",
    "torchvision==0.16.2+cu121",
    "torchaudio==2.1.2+cu121",
    "--extra-index-url https://download.pytorch.org/whl/cu121"
]

[project.urls]
Homepage = "https://github.com/enterprise/mlops-platform"
Documentation = "https://mlops-platform.readthedocs.io/"
Repository = "https://github.com/enterprise/mlops-platform"
"Bug Tracker" = "https://github.com/enterprise/mlops-platform/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src/*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  | \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pydantic.*",
    "sqlalchemy.*",
    "alembic.*",
    "boto3.*",
    "botocore.*",
    "celery.*",
    "neo4j.*",
    "mlflow.*",
    "ray.*",
    "dvc.*",
    "great_expectations.*",
    "whylogs.*",
    "evidently.*",
    "arize_phoenix.*",
    "feast.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/tests/*",
    "src/**/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
